+++ api/cm-cli/v1alpha1/doc.go
@@ -1,0 +1,6 @@
+// Copyright Contributors to the Open Cluster Management project
+
+// +k8s:deepcopy-gen=package
+// +kubebuilder:object:generate=true
+// +groupName=cm-cli.open-cluster-management.io
+package v1alpha1