+++ pkg/cmd/get/clusterpoolhosts/options.go
@@ -4,13 +4,13 @@ package clusterpoolhosts
 import (
 	genericclioptionscm "github.com/open-cluster-management/cm-cli/pkg/genericclioptions"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
+	"k8s.io/kubectl/pkg/cmd/get"
 )
 
 type Options struct {
 	//CMFlags: The generic optiosn from the cm cli-runtime.
-	CMFlags      *genericclioptionscm.CMFlags
-	NoHeaders    bool
-	OutputFormat string
+	CMFlags    *genericclioptionscm.CMFlags
+	PrintFlags *get.PrintFlags
 }
 
 func newOptions(cmFlags *genericclioptionscm.CMFlags, streams genericclioptions.IOStreams) *Options {